Build: use sbt-tpolecat plugin for stricter compilation
