+++ elastiknn-lucene/src/test/scala/com/klibisz/elastiknn/lucene/LuceneSupport.scala
@@ -32,6 +32,7 @@ trait LuceneSupport {
       } finally {
         indexWriter.close()
         tmpDir.delete()
+        ()
       }
     res
   }
