+++ project/plugins.sbt
@@ -1,1 +1,3 @@
 addSbtPlugin("pl.project13.scala" % "sbt-jmh" % "0.4.6")
+addSbtPlugin("org.typelevel" % "sbt-tpolecat" % "0.5.0")
+
